Restructure custom node; begin to style and add features
