+++ client/features/output-visualization-page/components/LinktaNode.tsx
@@ -1,8 +1,8 @@
-import type { ChangeEvent } from 'react';
-import React, { useState, memo } from 'react';
+import React, { memo } from 'react';
 import type { NodeProps } from 'reactflow';
 import { Handle, Position } from 'reactflow';
-import { Box, TextField } from '@mui/material';
+import { Box, Typography } from '@mui/material';
+import { DragIndicator, ExpandCircleDownOutlined } from '@mui/icons-material';
 import styles from '@styles/LinktaFlow.module.css';
 
 type LinktaNodeData = {
@@ -16,64 +16,34 @@ type LinktaNodeProps = NodeProps<LinktaNodeData> & {
 };
 
 const LinktaNode = memo(({ isConnectable, data }: LinktaNodeProps) => {
-  const [isFocused, setIsFocused] = useState(false);
-  const [label, setLabel] = useState(data.label || 'TCP/IP Model');
-
-  const onFocus = () => setIsFocused(true);
-  const onBlur = () => setIsFocused(false);
-
-  const handleLabelChange = (e: ChangeEvent<HTMLInputElement>) =>
-    setLabel(e.target.value);
+  const expandNode = () => {
+    console.log('NODE BUTTON');
+  };
 
   return (
-    <Box
-      className={`${styles.nodeWrapper} ${isFocused ? styles.nodeFocused : ''}`}
-    >
-      <Box className={styles.nodeColorBar} />
-      <Box className={styles.nodeContent}>
-        <TextField
-          hiddenLabel
-          value={label}
-          onFocus={onFocus}
-          onBlur={onBlur}
-          onChange={handleLabelChange}
-          size='small'
-          id='filled-hidden-label-small'
-          className={styles.textField}
-          InputProps={{
-            classes: {
-              input: styles.textFieldInput,
-              root: styles.textFieldFieldset,
-            },
-          }}
+    <Box className={`${styles.node}`}>
+      <Box className={styles.nodeGrabHandle}>
+        <DragIndicator className={`${styles.dragIndicator}`} />
+      </Box>
+      <Typography className={`${styles.nodeLabel}`}>{data.label}</Typography>
+      <Box className={`${styles.nodeButtonContainer}`}>
+        <ExpandCircleDownOutlined
+          onClick={expandNode}
+          className={`${styles.nodeButton}`}
         />
       </Box>
-      <Handle
-        type='source'
-        position={Position.Bottom}
-        id='a'
-        className={styles.handle}
-        isConnectable={isConnectable}
-      />
       <Handle
         type='target'
         position={Position.Top}
         id='d'
-        className={styles.handle}
+        className={styles.anchor}
         isConnectable={isConnectable}
       />
       <Handle
         type='source'
-        position={Position.Right}
-        id='b'
-        className={styles.handle}
-        isConnectable={isConnectable}
-      />
-      <Handle
-        type='target'
-        position={Position.Left}
-        id='c'
-        className={styles.handle}
+        position={Position.Bottom}
+        id='a'
+        className={styles.anchor}
         isConnectable={isConnectable}
       />
     </Box>
